backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
geoalchemy2==0.14.1
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
firebase-admin==6.2.0
twilio==8.10.2
python-multipart==0.0.6
cloudinary==1.36.0
boto3==1.28.85
scikit-learn==1.3.2
xgboost==2.0.2
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
slowapi==0.1.8
websockets==12.0
httpx==0.25.1
